+++ derive_rfc4226_wasm.go
